backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1

openai==1.30.1
langchain==0.2.0
langchain-community==0.2.0
langchain-openai==0.1.7
httpx==0.27.2
requests>=2.31
